Substituted EF Core with NHibernate

+++ Core/DTOs/AnimalDto.cs
@@ -5,7 +5,8 @@ public class AnimalDto
     public Guid Id { get; set; }
     public string? PetName { get; set; }
     public string? Kind { get; set; }
-    public int Age { get; set; }
+    public byte Age { get; set; }
+
     public Guid OwnerId { get; set; }
     public OwnerDto? Owner { get; set; }
 }
